Refactor: rename and replace java files
