+++ src/main/java/arrays/CreateArray.java
@@ -1,8 +1,8 @@
 package arrays;
 
-import classes.Bus;
-import classes.Student;
-import classes.User;
+import dataclasses.Bus;
+import dataclasses.Student;
+import dataclasses.User;
 
 public class CreateArray {
