+++ src/main/java/sortdir/comparators/StudentComparator.java
@@ -1,6 +1,6 @@
 package sortdir.comparators;
 
-import classes.Student;
+import dataclasses.Student;
 
 import java.util.Comparator;
