+++ src/main/java/sortdir/quicksorts/DataQuickSort.java
@@ -1,8 +1,8 @@
 package sortdir.quicksorts;
 
-import classes.Bus;
-import classes.Student;
-import classes.User;
+import dataclasses.Bus;
+import dataclasses.Student;
+import dataclasses.User;
 
 public class DataQuickSort<T> {
